github-issue-assistant/requirements.txt
# Core
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.3.0
python-multipart==0.0.6

# LLM Integration
openai>=1.40.0,<2.0
langchain==0.0.335
google-generativeai==0.8.0

# Frontend (if using Streamlit)
streamlit==1.27.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0
requests-mock==1.11.0
pytest-asyncio==0.21.1
requests==2.31.0

# Code quality
black==23.7.0
isort==5.12.0
flake8==6.1.0
